+++ client/components/AfterLogin/AfterLoginMobileContainer.js
@@ -1,0 +1,93 @@
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import { connect } from "react-redux";
+import {
+  Button,
+  Container,
+  Divider,
+  Grid,
+  Header,
+  Icon,
+  Image,
+  List,
+  Menu,
+  Responsive,
+  Segment,
+  Sidebar,
+  Visibility,
+  Input
+} from "semantic-ui-react";
+import { NavLink, withRouter, Link } from "react-router-dom";
+
+class MobileContainer extends Component {
+  state = {};
+
+  handlePusherClick = () => {
+    const { sidebarOpened } = this.state;
+
+    if (sidebarOpened) this.setState({ sidebarOpened: false });
+  };
+
+  handleToggle = () =>
+    this.setState({ sidebarOpened: !this.state.sidebarOpened });
+
+  render() {
+    const { children } = this.props;
+    const { sidebarOpened } = this.state;
+
+    return (
+      <Responsive {...Responsive.onlyMobile}>
+        <Sidebar.Pushable>
+          <Sidebar
+            as={Menu}
+            animation="uncover"
+            inverted
+            vertical
+            visible={sidebarOpened}
+          >
+            <Menu.Item as="a" active>
+              HOME
+            </Menu.Item>
+            <Menu.Item as="a">ADD ACCOUNTS</Menu.Item>
+            <Menu.Item as="a">PROFILE</Menu.Item>
+            <Menu.Item as="a">SETTING</Menu.Item>
+            <Menu.Item as="a">LOG OUT</Menu.Item>
+          </Sidebar>
+
+          <Sidebar.Pusher
+            dimmed={sidebarOpened}
+            onClick={this.handlePusherClick}
+            style={{ minHeight: "100vh" }}
+          >
+            <Segment
+              inverted
+              textAlign="center"
+              style={{ minHeight: 350, padding: "1em 0em" }}
+              vertical
+            >
+              <Container>
+                <Menu inverted pointing secondary size="large">
+                  <Menu.Item onClick={this.handleToggle}>
+                    <Icon name="sidebar" />
+                  </Menu.Item>
+                  <Menu.Item position="right">
+                    <Button as="a" inverted style={{ marginLeft: "0.5em" }}>
+                      LOG OUT
+                    </Button>
+                  </Menu.Item>
+                </Menu>
+              </Container>
+            </Segment>
+            {children}
+          </Sidebar.Pusher>
+        </Sidebar.Pushable>
+      </Responsive>
+    );
+  }
+}
+
+MobileContainer.propTypes = {
+  children: PropTypes.node
+};
+
+export default withRouter(MobileContainer);
