+++ package.json
@@ -45,7 +45,8 @@
     "semantic-react": "^0.12.0",
     "semantic-ui": "^2.3.1",
     "semantic-ui-react": "^0.80.0",
-    "sequelize": "^4.3.1"
+    "sequelize": "^4.3.1",
+    "victory": "^0.26.0"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0-beta.40",
